src/meter_ocr/requirements.txt
numpy==1.21.6
torch==1.11.0
#opencv-python-headless
opencv-python==4.5.5.64
pandas==1.3.5
requests==2.28.1
PyYAML==6.0
Pillow==9.1.0
tqdm==4.63.0
torchvision==0.12.0
matplotlib==3.5.1
seaborn==0.11.2
#transformers==4.27.2
#PyMuPDF==1.20.0
paddleocr
#pytesseract==0.3.9
paddlepaddle
#fastapi==0.89.1
#uvicorn==0.20.0
requests-toolbelt

#pip install paddleocr --upgrade
#pip install paddlepaddle
